Move metrics to own module
